third_party/WebKit/LayoutTests/html/dialog/modal-dialog-blocks-mouse-events-expected.txt

Test for bug 110952. Ensure that mouse events are not dispatched to an inert node. To test manually, move the mouse to the blue box, click, and then move the mouse outside. Then repeat for the red box. The test succeeds if both boxes turn green.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Clicking on inert box
PASS inertDivHandledEvent is false
PASS Object.keys(handledEvents.document).length is expectedEventCountForDocument
Clicking on non-inert box
PASS inertDivHandledEvent is false
PASS Object.keys(handledEvents.dialogDiv).length is events.length
PASS successfullyParsed is true

TEST COMPLETE
